+++ packages/video_player/video_player/android/src/test/java/io/flutter/plugins/videoplayer/VideoPlayerTest.java
@@ -10,6 +10,6 @@ public class VideoPlayerTest {
   // This is only a placeholder test and doesn't actually initialize the plugin.
   @Test
   public void initPluginDoesNotThrow() {
-    final VideoPlayerPlugin plugin = new VideoPlayerPlugin();
+    final VideoPlayerPluginCustom plugin = new VideoPlayerPluginCustom();
   }
 }