+++ packages/video_player/video_player/ios/Classes/FLTVideoPlayerPlugin.h
@@ -4,5 +4,5 @@
 
 #import <Flutter/Flutter.h>
 
-@interface FLTVideoPlayerPlugin : NSObject <FlutterPlugin>
+@interface FLTVideoPlayerPluginCustom : NSObject <FlutterPlugin>
 @end